backend/requirements.txt
fastapi==0.135.2
uvicorn==0.42.0
python-dotenv==1.2.2
google-genai==1.69.0
litellm==1.83.4
pydantic==2.12.5
Pillow==12.1.1
pyjwt==2.12.1
bcrypt==5.0.0
motor==3.7.1
redis==7.4.0
httpx==0.28.1
pydantic[email]==2.12.5
websockets==16.0
torchfree-ocr==1.1.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
